Added Tests, jacoco, deleted constructions from dto, translated all phrases from Russian into English

+++ src/main/java/com/sirotina/bankapp/entity/Account.java
@@ -41,7 +41,7 @@ public class Account {
     private AccountStatus status;
 
     @Column(name = "balance")
-    private Integer balance;
+    private BigDecimal balance;
 
     @Column(name = "currency_code")
     @Enumerated(EnumType.STRING)
